requirements.txt
fastapi
uvicorn
gunicorn
requests
python-dotenv
PyPDF2
pandas